Append a direct pointer to elts in uncached FcPattern
A FcPatternElt are either allocated in a cache that contains the
FcPattern owning it or allocated separately if it is not cached. When
a FcPatternElt is allocated separately, runtime memory leak checks such
as LeakSanitizer cannot understand this reference because it's not a
normal pointer, and reports false leaks. Such false reports are very
noisy when applying runtime memory checks to an application that uses
Fontconfig.

Append a direct pointer to elts to aid such runtime checks.

Signed-off-by: Akihiko Odaki <akihiko.odaki@gmail.com>